pyproject.toml
[project]
name = "atrun"
version = "0.13.0"
description = "Social package distribution on AT Protocol"
requires-python = ">=3.12"
license = "Apache-2.0"
dependencies = [
    "click>=8.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.urls]
Repository = "https://github.com/mhalle/atrun"

[project.scripts]
atrun = "atrun.cli:cli"

[dependency-groups]
dev = ["pytest>=8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
